Fixed mongo and mysql with latest e2e changes
