+++ tests/e2e/tests/add-mysql-datasource-flow.spec.ts
@@ -1,98 +1,17 @@
-import { type ConsoleMessage, type Page, test, expect } from '@playwright/test';
-
-// Helper function to navigate to MySQL data source form
-async function navigateToMySQLForm(page: Page) {
-  console.log('🧭 Navigating to application...');
-  await page.goto('/');
-
-  // Wait for the page to load
-  await page.waitForLoadState('networkidle');
-  console.log('✅ Page loaded successfully');
-
-  try {
-    console.log('🔍 Checking for telemetry consent page...');
-
-    const telemetryHeading = page.locator('h1:has-text("Help us improve liblab ai")');
-    await telemetryHeading.waitFor({ state: 'visible', timeout: 5000 });
-    console.log('📋 Found telemetry consent page, clicking Decline...');
-
-    const declineButton = page.locator('button:has-text("Decline")');
-    await declineButton.waitFor({ state: 'visible' });
-    await declineButton.click();
-
-    await page.waitForLoadState('networkidle');
-    console.log('✅ Declined telemetry, waiting for redirect...');
-  } catch {
-    console.warn('ℹ️ No telemetry consent page found, continuing...');
-  }
-
-  try {
-    console.log('🔍 Checking for data source connection page...');
-
-    const dataSourceHeading = page.locator('h1:has-text("Let\'s connect your data source")');
-    await dataSourceHeading.waitFor({ state: 'visible', timeout: 5000 });
-    console.log('💾 Found data source connection page, connecting to sample database...');
-
-    const connectButton = page.locator('button:has-text("Connect")');
-    await connectButton.waitFor({ state: 'visible', timeout: 10000 });
-    console.log('🔗 Found Connect button, clicking...');
-    await connectButton.click();
-
-    await page.waitForLoadState('networkidle');
-    console.log('✅ Connected to sample database, waiting for redirect...');
-  } catch {
-    console.warn('ℹ️ No data source connection page found, continuing...');
-  }
-
-  console.log('🔍 Looking for settings button...');
-
-  const settingsButton = page.locator('[data-testid="settings-button"]');
-  await settingsButton.waitFor({ state: 'attached', timeout: 10000 });
-
-  // Use JavaScript click instead of Playwright click to bypass viewport restrictions
-  await page.evaluate(() => {
-    const button = document.querySelector('[data-testid="settings-button"]') as HTMLElement;
-
-    if (button) {
-      button.click();
-    } else {
-      throw new Error('Settings button not found in DOM');
-    }
-  });
-  console.log('✅ Successfully clicked settings button');
-
-  await page.getByRole('button', { name: 'Add Data Source' }).click();
-
-  // Look for the database type selector
-  console.log('🔍 Looking for database type selector...');
-
-  const dbTypeSelector = page.locator('select, [data-testid="add-data-source-select"]');
-  await dbTypeSelector.waitFor({ state: 'visible', timeout: 10000 });
-  console.log('✅ Found database type selector, selecting MySQL...');
-
-  // Click to open the dropdown and select MySQL
-  await dbTypeSelector.click();
-
-  await page.waitForLoadState('domcontentloaded');
-
-  const mysqlOption = page.locator('[id="mysql"]');
-  await mysqlOption.waitFor({ state: 'visible', timeout: 5000 });
-  await mysqlOption.click();
-  console.log('✅ Selected MySQL database type');
-}
+import { type Page, test, expect } from '@playwright/test';
+import { performInitialSetup, navigateToSettings, navigateToDataSourceForm } from '../helpers/setup';
 
 test.describe('Add MySQL Data Source Flow', () => {
+  test.beforeEach(async ({ page }) => {
+    test.setTimeout(120000);
+    await performInitialSetup(page);
+    await navigateToSettings(page);
+  });
+
   test('Create MySQL data source with valid connection string', async ({ page }: { page: Page }) => {
-    test.setTimeout(120000); // 2 minutes for this specific test
-
-    // Enable browser console logging for debugging
-    page.on('console', (msg: ConsoleMessage) => console.log('🖥️ Browser console:', msg.text()));
-    page.on('pageerror', (error: Error) => console.log('🖥️ Browser error:', error.message));
-
     console.log('Starting MySQL data source creation test...');
 
-    // Use helper function to navigate to MySQL form
-    await navigateToMySQLForm(page);
+    await navigateToDataSourceForm(page, 'mysql');
 
     console.log('🔍 Looking for database name input...');
 
@@ -150,16 +69,9 @@ test.describe('Add MySQL Data Source Flow', () => {
   });
 
   test('Validate MySQL connection string format and required fields', async ({ page }: { page: Page }) => {
-    test.setTimeout(120000); // 2 minutes for this specific test
-
-    // Enable browser console logging for debugging
-    page.on('console', (msg: ConsoleMessage) => console.log('🖥️ Browser console:', msg.text()));
-    page.on('pageerror', (error: Error) => console.log('🖥️ Browser error:', error.message));
-
     console.log('Starting MySQL validation test...');
 
-    // Use helper function to navigate to MySQL form
-    await navigateToMySQLForm(page);
+    await navigateToDataSourceForm(page, 'mysql');
 
     // Test 1: Try to create without filling required fields
     console.log('🔍 Testing form validation - attempting to create without required fields...');
